# Speed up `imports:associateProfile` task