--- a/core/src/modules/ops/export.ts
+++ b/core/src/modules/ops/export.ts
@@ -91,7 +91,6 @@ export namespace ExportOps {
           [Op.or]: [null, { [Op.lt]: new Date().getTime() - delayMs }],
         },
       },
-      order: [["sendAt", "asc"]],
       limit,
     });