--- a/core/src/modules/ops/import.ts
+++ b/core/src/modules/ops/import.ts
@@ -29,7 +29,6 @@ export namespace ImportOps {
         },
       },
       limit,
-      order: [["createdAt", "asc"]],
     });
 
     if (imports.length > 0) {