--- a/core/src/modules/ops/profileProperty.ts
+++ b/core/src/modules/ops/profileProperty.ts
@@ -102,7 +102,6 @@ export namespace ProfilePropertyOps {
           [Op.or]: [null, { [Op.lt]: new Date().getTime() - delayMs }],
         },
       },
-      order: [["stateChangedAt", "ASC"]],
       limit,
     });