## What's Inside:

1. Add an index to the `profileProperties` table on `["propertyId", "rawValue"]`- this speeds up the `import:associateProfile` task significantly (`findOrCreateByUniqueProfileProperties`), as it makes finding an existing matching profileProperty faster. 
2. Remove a 'double-set' of OldProfileProperties and OldGroups when updating an import (we were doing it twice)
3. Remove the SORT from our batch tasks (eg `import:associateProfiles` -> `processPendingImportsForAssociation`).  This was the biggest source of the slowdown in finding which Imports, ProfileProperties or Exports to move along in the next batch.
   * **Side Effect** -> This means that the first imported Profiles will likely not be the first exported.  The order will be more random.  With Exports, this delays even more the the time between when they were calculated from when they will be sent 
4. `createNullProperties` now makes a single insert statement, and writes all the logs as one insert statement too
4. Non-CLS Tasks are now instrumented with APM (Sentry)

## Ideas:
* remove locks
* do things in batches
* pre-load things
* remove redundant things
* Indices in the DB?

## Investigation Notes:

Disabling the locks (`waitForLock`) in the `findOrCreateByUniqueProfileProperties` method did not make a noticeable speed difference.

---

Confirmed that it is NOT faster to 2 to queries than one query with an OR (at least in this case):

```sql
--- 1 Query ---
EXPLAIN ANALYZE SELECT * FROM "profileProperties" AS "ProfileProperty" WHERE "ProfileProperty"."propertyId" = 'email' AND "ProfileProperty"."state" = 'pending' AND ("ProfileProperty"."startedAt" IS NULL OR "ProfileProperty"."startedAt" < '2021-05-22 18:38:00.182 +00:00') ORDER BY "ProfileProperty"."stateChangedAt" ASC LIMIT 500;
-- 40ms

--- 2 queries ---
EXPLAIN ANALYZE SELECT * FROM "profileProperties" AS "ProfileProperty" WHERE "ProfileProperty"."propertyId" = 'email' AND "ProfileProperty"."state" = 'pending' AND ( "ProfileProperty"."startedAt" < '2021-05-22 18:38:00.182 +00:00') ORDER BY "ProfileProperty"."stateChangedAt" ASC LIMIT 500;
-- 34ms
-- + 
EXPLAIN ANALYZE SELECT * FROM "profileProperties" AS "ProfileProperty" WHERE "ProfileProperty"."propertyId" = 'email' AND "ProfileProperty"."state" = 'pending' AND ("ProfileProperty"."startedAt" IS NULL) ORDER BY "ProfileProperty"."stateChangedAt" ASC LIMIT 500;
-- 40ms
-- = 74ms total
```

On the other hand, the ORDER BY was the thing causing the biggest slowdown

```sql
EXPLAIN ANALYZE 
SELECT * FROM "profileProperties" AS "ProfileProperty" 
WHERE 
	"ProfileProperty"."propertyId" = 'email' 
	AND "ProfileProperty"."state" = 'pending' 
	AND ("ProfileProperty"."startedAt" IS NULL OR "ProfileProperty"."startedAt" < '2021-05-22 18:38:00.182 +00:00') 
LIMIT 500;
-- 2ms
```

---