@evantahler 2021-05-21
A test @bleonard ran with a large dataset:

#### Before the Migration

```
2021-05-21T22:47:49.549Z - info: [ worker ] task success workerId=4 class=import:associateProfile queue=imports args={"importId":"imp_83005ff6-f8b9-43c8-ac90-f03f15a6e863"} duration=1397 
2021-05-21T22:47:50.308Z - info: [ worker ] task success workerId=4 class=import:associateProfile queue=imports args={"importId":"imp_1fc229fd-362c-467a-af94-bdb0e0d92316"} duration=759 
2021-05-21T22:47:50.418Z - info: [ worker ] task success workerId=1 class=import:associateProfile queue=imports args={"importId":"imp_d47974ab-163d-476a-9e7a-71460014a0cd"} duration=2301 
2021-05-21T22:47:50.488Z - info: [ worker ] task success workerId=3 class=import:associateProfile queue=imports args={"importId":"imp_6e161376-4780-4b82-81b0-b1e13bb6586b"} duration=1593 
2021-05-21T22:47:50.642Z - info: [ worker ] task success workerId=2 class=import:associateProfile queue=imports args={"importId":"imp_104e4851-5e62-4702-807a-af83e520c232"} duration=2247 
2021-05-21T22:47:50.847Z - info: [ worker ] task success workerId=4 class=import:associateProfile queue=imports args={"importId":"imp_ea0263a3-012a-4031-936c-d342dcda760d"} duration=537 
2021-05-21T22:47:51.485Z - info: [ worker ] task success workerId=4 class=import:associateProfile queue=imports args={"importId":"imp_aa55ec6f-2c49-44a2-b98c-294f0bc82526"} duration=638 
2021-05-21T22:47:51.757Z - info: [ worker ] task success workerId=1 class=import:associateProfile queue=imports args={"importId":"imp_27cf7d03-8ea0-4a58-bde9-853dd032c7c4"} duration=1338 
2021-05-21T22:47:52.308Z - info: [ worker ] task success workerId=1 class=import:associateProfile queue=imports args={"importId":"imp_89d3dde8-cea2-4c41-9970-fc6205cefc44"} duration=550 
2021-05-21T22:47:52.699Z - info: [ worker ] task success workerId=3 class=import:associateProfile queue=imports args={"importId":"imp_456c3b0a-8b9e-4f2a-b3fa-2756d858d6fc"} duration=2209 
2021-05-21T22:47:52.842Z - info: [ worker ] task success workerId=1 class=import:associateProfile queue=imports args={"importId":"imp_aa914fb7-5e9b-4d0e-bec1-836e246e332b"} duration=532 
2021-05-21T22:47:52.872Z - info: [ worker ] task success workerId=2 class=import:associateProfile queue=imports args={"importId":"imp_51eb2a7b-1524-4fda-8d2f-54fafa1f45fd"} duration=2229 
2021-05-21T22:47:53.403Z - info: [ worker ] task success workerId=1 class=import:associateProfile queue=imports args={"importId":"imp_71a29b35-5641-4a99-bd30-56ec1aa43a49"} duration=561 
```

#### After the Migration

```
2021-05-21T22:49:42.983Z - info: [ worker ] task success workerId=4 class=import:associateProfile queue=imports args={"importId":"imp_c75a71e4-1e4c-4de7-b576-ab909a1edbe8"} duration=258 
2021-05-21T22:49:43.206Z - info: [ worker ] task success workerId=1 class=import:associateProfile queue=imports args={"importId":"imp_3a5a0cce-43c3-4893-9dc3-34cccf956105"} duration=276 
2021-05-21T22:49:43.236Z - info: [ worker ] task success workerId=4 class=import:associateProfile queue=imports args={"importId":"imp_8038e2f5-63b7-4a36-9ada-fd22cd395240"} duration=252 
2021-05-21T22:49:43.236Z - info: [ worker ] task success workerId=2 class=import:associateProfile queue=imports args={"importId":"imp_f779afef-7a0e-4a1c-83bf-ffdc26afef13"} duration=252 
2021-05-21T22:49:43.321Z - info: [ worker ] task success workerId=3 class=import:associateProfile queue=imports args={"importId":"imp_ce5333bd-f7f0-4968-b517-cdea8ed69528"} duration=360 
2021-05-21T22:49:43.498Z - info: [ worker ] task success workerId=4 class=import:associateProfile queue=imports args={"importId":"imp_2ce2834e-6b19-4b09-8515-3be0bf26fea1"} duration=262 
2021-05-21T22:49:43.498Z - info: [ worker ] task success workerId=2 class=import:associateProfile queue=imports args={"importId":"imp_a308a701-10b7-461a-934c-fa39245f6b90"} duration=261 
2021-05-21T22:49:43.500Z - info: [ worker ] task success workerId=1 class=import:associateProfile queue=imports args={"importId":"imp_891703fe-1f73-4c1c-a610-c362b5130009"} duration=294 
2021-05-21T22:49:43.579Z - info: [ worker ] task success workerId=3 class=import:associateProfile queue=imports args={"importId":"imp_e9d169f0-97aa-4f92-954f-2618b2082eb4"} duration=257 
2021-05-21T22:49:43.751Z - info: [ worker ] task success workerId=2 class=import:associateProfile queue=imports args={"importId":"imp_ef01b5c0-ec0f-4ea0-9ca7-44be1f61779f"} duration=252 
2021-05-21T22:49:43.754Z - info: [ worker ] task success workerId=1 class=import:associateProfile queue=imports args={"importId":"imp_a38d65e1-d16a-47e2-a649-75e8b45e9339"} duration=252 
```